+++ src/pages/Shops/Shop/ShopInvoice.jsx
@@ -1,52 +1,52 @@
-import React, { useEffect, useState } from 'react';
-import MyTable from '../../../components/MyTable';
-import PageLayout from '../../../components/PageLayout';
-import MyDialog from '../../../components/MyDialog';
-import Box from '@mui/material/Box';
-import Breadcrumbs from '@mui/material/Breadcrumbs';
-import Button from '@mui/material/Button';
-import Typography from '@mui/material/Typography';
-import useTable from '../../../hooks/useTable';
-import { useSelector } from 'react-redux';
-import * as Yup from 'yup';
+import React, { useEffect, useState } from "react";
+import MyTable from "../../../components/MyTable";
+import PageLayout from "../../../components/PageLayout";
+import MyDialog from "../../../components/MyDialog";
+import Box from "@mui/material/Box";
+import Breadcrumbs from "@mui/material/Breadcrumbs";
+import Button from "@mui/material/Button";
+import Typography from "@mui/material/Typography";
+import useTable from "../../../hooks/useTable";
+import { useSelector } from "react-redux";
+import * as Yup from "yup";
 // import { addCustomer, deleteCustomer, editCustomer } from '../store/users';
-import { useDispatch } from 'react-redux';
+import { useDispatch } from "react-redux";
 import {
   addProduct,
   deleteProduct,
   editProduct,
   getProducts,
-} from '../../../store/products';
-import Logs from '../../../components/Logs';
-import PageHeading from '../../../components/PageHeading';
-import TableSubHeader from '../../../components/TableSubHeader';
-import LogsTable from '../../../components/LogsTable';
-import { Link, useParams } from 'react-router-dom';
+} from "../../../store/products";
+import Logs from "../../../components/Logs";
+import PageHeading from "../../../components/PageHeading";
+import TableSubHeader from "../../../components/TableSubHeader";
+import LogsTable from "../../../components/LogsTable";
+import { Link, useParams } from "react-router-dom";
 import {
   addBillItem,
   deleteBillItem,
   editBillItem,
   getBill,
-} from '../../../store/bill';
-import Grid from '@mui/material/Grid';
-const PAGE_TITLE = 'فاتورة رقم';
+} from "../../../store/bill";
+import Grid from "@mui/material/Grid";
+const PAGE_TITLE = "فاتورة رقم";
 
 const DAILY_ROW_INTIAL_VALUE = {
-  id: '',
-  ProductId: '',
+  id: "",
+  ProductId: "",
   count: 0,
   value: 0,
   values: 0,
   // totalValue: '',
   // totalValues: '',
-  note: '',
+  note: "",
 };
 
 const validationSchema = Yup.object({
-  ProductId: Yup.number().required('يرجى إدخال اسم المنتج'),
-  count: Yup.number().required('يرجى إدخال الكمية المتاحة'),
-  value: Yup.number().required('يرجى إدخال مبلغ القطعة الواحدة'),
-  values: Yup.number().required('يرجى إدخال قيمة القطعة الواحدة'),
+  ProductId: Yup.number().required("يرجى إدخال اسم المنتج"),
+  count: Yup.number().required("يرجى إدخال الكمية المتاحة"),
+  value: Yup.number().required("يرجى إدخال مبلغ القطعة الواحدة"),
+  values: Yup.number().required("يرجى إدخال قيمة القطعة الواحدة"),
   note: Yup.string(),
 });
 
@@ -61,8 +61,9 @@ const CustomerInvoice = () => {
     ...billItem,
     ProductId: billItem.ProductId,
     name: billItem.Product.name,
-    totalValue: parseInt(billItem.value, 10) * parseInt(billItem.count, 10),
-    totalValues: parseInt(billItem.values, 10) * parseInt(billItem.count, 10),
+    totalValue: parseFloat(billItem.value, 10) * parseFloat(billItem.count, 10),
+    totalValues:
+      parseFloat(billItem.values, 10) * parseFloat(billItem.count, 10),
   }));
   const productsRows = useSelector((state) => state.products.data);
   const products = productsRows.map((row) => ({
@@ -72,75 +73,75 @@ const CustomerInvoice = () => {
   const COLUMNS = [
     {
       id: `ProductId`,
-      label: 'رقم المنتج',
+      label: "رقم المنتج",
       minWidth: 80,
-      align: 'center',
+      align: "center",
       options: products,
       isField: true,
     },
     {
       id: `id`,
-      label: 'رقم سطر الفاتورة',
+      label: "رقم سطر الفاتورة",
       minWidth: 80,
-      align: 'center',
+      align: "center",
       isField: false,
     },
     {
-      id: 'name',
-      label: 'اسم المنتج',
+      id: "name",
+      label: "اسم المنتج",
       minWidth: 150,
-      align: 'left',
+      align: "left",
       required: true,
     },
     {
-      id: 'count',
-      label: 'العدد',
+      id: "count",
+      label: "العدد",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: true,
       required: true,
     },
     {
-      id: 'value',
-      label: 'المبلغ للواحدة',
+      id: "value",
+      label: "المبلغ للواحدة",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: true,
     },
     {
-      id: 'values',
-      label: 'القيمة للواحدة',
+      id: "values",
+      label: "القيمة للواحدة",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: true,
     },
     {
-      id: 'totalValue',
-      label: 'المبلغ الاجمالي',
+      id: "totalValue",
+      label: "المبلغ الاجمالي",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: false,
     },
     {
-      id: 'totalValues',
-      label: 'القيمة الاجمالية',
+      id: "totalValues",
+      label: "القيمة الاجمالية",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: false,
     },
 
     {
-      id: 'note',
-      label: 'ملاحظة',
+      id: "note",
+      label: "ملاحظة",
       minWidth: 100,
-      align: 'center',
+      align: "center",
       isField: true,
     },
   ];
 
   const dispatchers = {
     add: (row) => {
-      console.log(row, 'test');
+      console.log(row, "test");
       dispatch(addBillItem({ billId: invoiceId, row, shop: true })).then(() =>
         dispatch(getBill(invoiceId))
       );
@@ -168,65 +169,65 @@ const CustomerInvoice = () => {
         <Grid
           item
           xs={12}
-          sx={{ display: 'flex', justifyContent: 'space-between' }}
+          sx={{ display: "flex", justifyContent: "space-between" }}
         >
           <Grid container>
             <Grid
               item
               xs={6}
-              sx={{ display: 'flex', justifyContent: 'space-between' }}
+              sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Typography variant="body1">ملاحظة الفاتورة</Typography>
-              <Typography variant="body1">{bill.note || 'لايوجد'}</Typography>
+              <Typography variant="body1">{bill.note || "لايوجد"}</Typography>
             </Grid>
           </Grid>
         </Grid>
         <Grid
           item
           xs={12}
-          sx={{ display: 'flex', justifyContent: 'space-between' }}
+          sx={{ display: "flex", justifyContent: "space-between" }}
         >
           <Grid container>
             <Grid
               item
               xs={6}
-              sx={{ display: 'flex', justifyContent: 'space-between' }}
+              sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Typography variant="body1">اجمالي المبلغ</Typography>
-              <Typography variant="body1">{bill.value || 'لايوجد'}</Typography>
+              <Typography variant="body1">{bill.value || "لايوجد"}</Typography>
             </Grid>
           </Grid>
         </Grid>
         <Grid
           item
           xs={12}
-          sx={{ display: 'flex', justifyContent: 'space-between' }}
+          sx={{ display: "flex", justifyContent: "space-between" }}
         >
           <Grid container>
             <Grid
               item
               xs={6}
-              sx={{ display: 'flex', justifyContent: 'space-between' }}
+              sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Typography variant="body1">اجمالي القيمة</Typography>
-              <Typography variant="body1">{bill.values || 'لايوجد'}</Typography>
+              <Typography variant="body1">{bill.values || "لايوجد"}</Typography>
             </Grid>
           </Grid>
         </Grid>
         <Grid
           item
           xs={12}
-          sx={{ display: 'flex', justifyContent: 'space-between' }}
+          sx={{ display: "flex", justifyContent: "space-between" }}
         >
           <Grid container>
             <Grid
               item
               xs={6}
-              sx={{ display: 'flex', justifyContent: 'space-between' }}
+              sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Typography variant="body1">نوع الفاتورة</Typography>
               <Typography variant="body1">
-                {bill.billType || 'لايوجد'}
+                {bill.billType || "لايوجد"}
               </Typography>
             </Grid>
           </Grid>
@@ -234,17 +235,17 @@ const CustomerInvoice = () => {
         <Grid
           item
           xs={12}
-          sx={{ display: 'flex', justifyContent: 'space-between' }}
+          sx={{ display: "flex", justifyContent: "space-between" }}
         >
           <Grid container>
             <Grid
               item
               xs={6}
-              sx={{ display: 'flex', justifyContent: 'space-between' }}
+              sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Typography variant="body1">تاريخ الاصدار</Typography>
               <Typography variant="body1">
-                {bill.createdAt || 'لايوجد'}
+                {bill.createdAt || "لايوجد"}
               </Typography>
             </Grid>
           </Grid>
@@ -260,19 +261,19 @@ const CustomerInvoice = () => {
           validationSchema={validationSchema}
           dispatchers={dispatchers}
         />
-        {bill.billType === 'ادخال' && (
+        {bill.billType === "ادخال" && (
           <Button
             variant="contained"
             sx={{
-              height: '60px',
-              margin: '2rem auto',
-              display: 'block',
-              width: '50%',
+              height: "60px",
+              margin: "2rem auto",
+              display: "block",
+              width: "50%",
             }}
             color="error"
             onClick={handleOpenAddDialog}
           >
-            {'إضافة منتج للفاتورة'}
+            {"إضافة منتج للفاتورة"}
           </Button>
         )}
       </Box>
